feat: add support for printing HTML document
The CUPS print service does not support printing HTML documents, so we can't simply send it HTML like we do with PDF documents. Instead, we convert the HTML to PDF before printing it using Google Chrome's headless `--print-to-pdf` option.

The problem with this approach is that we can't just grab the HTML from the visible running Chrome and send it to the headless Chrome. That actually does get us a surprising amount of the way there, but any remotely loaded assets do not get included. To fix this, we modify the HTML before sending it to the headless Chrome by adding the venerable `base` tag and like so: `<base href="http://localhost:3000">`, with the `href` being whatever the `origin` of the `Referer` header is. Once that tag is there Chrome can find the assets again and download them as normal.